Refactor infassetinfo_tables.py
