--- src/zepben/evolve/database/sqlite/tables/iec61968/infiec61968/infassetinfo/__init__.py
@@ -1,5 +1,0 @@
-#  Copyright 2023 Zeppelin Bend Pty Ltd
-#
-#  This Source Code Form is subject to the terms of the Mozilla Public
-#  License, v. 2.0. If a copy of the MPL was not distributed with this
-#  file, You can obtain one at https://mozilla.org/MPL/2.0/.
